+++ src/wallet/wallet.js
@@ -18,8 +18,10 @@ export class Wallet {
    * @param {Keypair|string} keypair User's key pair or a secret seed.
    * @param {string} accountId User's account ID.
    * @param {string} [walletId] Wallet ID.
+   * @param {string} [sessionId] Session ID.
+   * @param {string} [sessionKey] Session key.
    */
-  constructor (email, keypair, accountId, walletId) {
+  constructor (email, keypair, accountId, walletId, sessionId, sessionKey) {
     if (isNil(email)) {
       throw new Error('Email is required.')
     }
@@ -43,10 +45,20 @@ export class Wallet {
       throw new Error('Hex encoded wallet ID expected.')
     }
 
+    if (sessionId && !isString(sessionId)) {
+      throw new Error('Hex encoded session ID expected.')
+    }
+
+    if (sessionKey && !isString(sessionKey)) {
+      throw new Error('Hex encoded session key expected.')
+    }
+
     this._email = email
     this._keypair = keypair
     this._accountId = accountId
     this._id = walletId
+    this._sessionId = sessionId
+    this._sessionKey = sessionKey
   }
 
   /**
@@ -77,6 +89,8 @@ export class Wallet {
    * @param {string} opts.salt Salt used for encryption.
    * @param {string} opts.email User's email.
    * @param {string} opts.password User's password.
+   * @param {string} opts.sessionId Session id.
+   * @param {string} opts.sessionKey Session key.
    * @param {string} [opts.accountId] User's account ID.
    */
   static fromEncrypted (opts) {
@@ -96,7 +110,9 @@ export class Wallet {
       opts.email,
       Keypair.fromSecret(decryptedKeychain.seed),
       opts.accountId || Keypair.fromSecret(decryptedKeychain.seed).accountId(),
-      sjcl.codec.hex.fromBits(rawWalletId)
+      sjcl.codec.hex.fromBits(rawWalletId),
+      opts.sessionId,
+      opts.sessionKey
     )
   }
 
@@ -174,6 +190,20 @@ export class Wallet {
    */
   get keypair () {
     return this._keypair
+  }
+
+  /**
+   * Get session ID.
+   */
+  get sessionId () {
+    return this._sessionId
+  }
+
+  /**
+   * Get session key.
+   */
+  get sessionKey () {
+    return this._sessionKey
   }
 
   /**